experiments/Data Extraction/Manual Extraction/test #2/requirements.txt
# PDF Expense Extractor Dependencies
# Core dependencies for PDF processing and OCR

# PDF to image conversion
pdf2image==1.17.0

# OCR text extraction
pytesseract==0.3.10

# Table formatting for CLI output
tabulate==0.9.0

# Additional dependencies that may be needed:
# Pillow is usually installed with pdf2image but listing explicitly
Pillow>=10.0.0

# Note: You also need to install Tesseract OCR on your system:
# 
# Windows:
#   Download from: https://github.com/UB-Mannheim/tesseract/wiki
#   Or use: winget install UB-Mannheim.TesseractOCR
#
# macOS:
#   brew install tesseract
#
# Ubuntu/Debian:
#   sudo apt-get install tesseract-ocr
#
# CentOS/RHEL:
#   sudo yum install tesseract